// api/src/types/featureTypes.ts
// Define the feature types that can be tracked
export enum FeatureType {
  CHAT = 'chat',
  TEXT_TO_SPEECH = 'textToSpeech',
  SPEECH_TO_TEXT = 'speechToText',
  IMAGE_GENERATION = 'imageGeneration',
  DOCUMENT_ANALYSIS = 'documentAnalysis',
  DOCUMENT_PROCESSING = 'documentProcessing',
  CONVERSATION = 'conversation',
  DEEP_LEARNING = 'deepLearning',
  NLP = 'nlp'
}